# openapi.yaml
openapi: 3.1.0
info:
  title: Self Agent ID API
  version: 1.0.0
  description: >
    On-chain AI agent identity registry with proof-of-human verification.
    Register, deregister, and query agents — no API keys required.
    Sessions use encrypted tokens with a 30-minute TTL.
  contact:
    name: Self Protocol
    url: https://github.com/selfxyz/self-agent-id
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://self-agent-id.vercel.app
    description: Production
...
security: []
...
components:
  securitySchemes:
    selfAgentAuth:
      type: apiKey
      in: header
      name: x-self-agent-signature
      description: >
        ECDSA signature produced by `SelfAgent.signRequest()`.
        Must be accompanied by `x-self-agent-address` and
        `x-self-agent-timestamp` headers.
    aaProxyToken:
      type: apiKey
      in: header
      name: x-aa-proxy-token
      description: >
        Short-lived token issued by POST /api/aa/token.
        Required for bundler and paymaster proxy endpoints.
  parameters:
    chainIdPath:
      name: chainId
      in: path
      required: true
      schema:
        type: integer
        enum: [42220, 11142220]
      description: "42220 (Celo Mainnet) or 11142220 (Celo Sepolia Testnet)"
    agentIdPath:
      name: agentId
      in: path
      required: true
      schema:
        type: integer
      description: On-chain agent token ID
...
    chainIdQuery:
      name: chainId
      in: query
      required: true
      schema:
        type: string
        enum: ["42220", "11142220"]
      description: "42220 (mainnet) or 11142220 (testnet)"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    VerifyResponse:
      type: object
      properties:
        agentId:
          type: integer
        isVerified:
          type: boolean
        proofProvider:
          type: string
        strengthLabel:
          type: string
        humanAgentCount:
          type: integer
        maxAgentsPerHuman:
          type: integer
    AgentCardResponse:
      type: object
      description: A2A-compatible agent identity card (JSON-LD)
      additionalProperties: true
...
    CensusPostResponse:
      type: object
      properties:
        recorded:
          type: boolean
        totalAgents:
          type: integer
        yourEntry:
          type: object
          properties:
            agentAddress:
              type: string
            agentId:
              type: string
            nationality:
              type: string
            olderThan:
              type: integer
            ofac:
              type: array
              items:
                type: boolean
            timestamp:
              type: integer
    CensusGetResponse:
      type: object
      properties:
        topCountries:
          type: array
          items:
            type: object
            properties:
              country:
                type: string
              count:
                type: integer
        verifiedOver18:
          type: integer
        verifiedOver21:
          type: integer
        ofacClear:
          type: integer
        totalAgents:
          type: integer
...
    AaTokenResponse:
      type: object
      properties:
        token:
          type: string
        expiresAt:
          type: integer
          description: Unix timestamp (ms)
    JsonRpcRequest:
      type: object
      required: [jsonrpc, method]
      properties:
        jsonrpc:
          type: string
          enum: ["2.0"]
        method:
          type: string
        params:
          type: array
          items: {}
        id:
          oneOf:
            - type: string
            - type: integer
            - type: "null"
    JsonRpcResponse:
      type: object
      description: Standard JSON-RPC 2.0 response from upstream bundler/paymaster
      additionalProperties: true

# ─── Paths ───────────────────────────────────────────────────────────
paths:
...
  /api/agent/verify/{chainId}/{agentId}:
    get:
      tags: [Query]
      summary: Verify agent proof-of-human
      operationId: verifyAgent
      description: >
        Checks whether an agent has valid proof-of-human verification,
        the proof provider address, verification strength label, and
        Sybil metrics.
      parameters:
        - $ref: "#/components/parameters/chainIdPath"
        - $ref: "#/components/parameters/agentIdPath"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
              example:
                agentId: 5
                isVerified: true
                proofProvider: "0x5E61c3051Bf4115F90AacEAE6212bc419f8aBB6c"
                strengthLabel: Standard
                humanAgentCount: 1
                maxAgentsPerHuman: 1
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Discovery ────────────────────────────────────────────────────
  /api/cards/{chainId}/{agentId}:
    get:
      tags: [Discovery]
      summary: Get agent A2A card
      operationId: getAgentCard
      description: >
        Returns the A2A-compatible agent identity card stored on-chain
        as JSON metadata.
      parameters:
        - $ref: "#/components/parameters/chainIdPath"
        - $ref: "#/components/parameters/agentIdPath"
      responses:
        "200":
          description: Agent card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentCardResponse"
        "404":
          description: Agent not found or no card set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/demo/census:
    post:
      tags: [Demo]
      summary: Contribute credentials to census
      operationId: postCensus
      description: >
        Agent-gated data contribution. Verified agent credentials
        (nationality, age, OFAC) are recorded in the census store.
      security:
        - selfAgentAuth: []
      responses:
        "200":
          description: Credentials recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CensusPostResponse"
        "403":
          description: Agent verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [Demo]
      summary: Read census aggregate stats
      operationId: getCensus
      description: >
        Agent-gated read of aggregate census data. Returns top countries,
        age distribution, and OFAC screening stats.
      security:
        - selfAgentAuth: []
      responses:
        "200":
          description: Aggregate census stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CensusGetResponse"
              example:
                topCountries:
                  - country: GBR
                    count: 12
                  - country: USA
                    count: 8
                  - country: DEU
                    count: 5
                verifiedOver18: 20
                verifiedOver21: 15
                ofacClear: 25
                totalAgents: 25
        "403":
          description: Agent verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/aa/token:
    post:
      tags: [AA Proxy]
      summary: Issue AA proxy token
      operationId: issueAaToken
      description: >
        Issues a short-lived token for accessing the bundler and
        paymaster proxy endpoints. Origin-restricted and rate-limited
        (30 req/min per IP).
      parameters:
        - $ref: "#/components/parameters/chainIdQuery"
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AaTokenResponse"
              example:
                token: "eyJhbGciOi..."
                expiresAt: 1740001800000
        "400":
          description: Unsupported or missing chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Origin check failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  retryAfterMs:
                    type: integer
  /api/aa/bundler:
    post:
      tags: [AA Proxy]
      summary: Bundler JSON-RPC proxy
      operationId: bundlerProxy
      description: >
        Proxies ERC-4337 bundler JSON-RPC calls to Pimlico. Requires an
        `x-aa-proxy-token` header. Allowed methods:
        `eth_chainId`, `eth_supportedEntryPoints`,
        `eth_estimateUserOperationGas`, `eth_sendUserOperation`,
        `eth_getUserOperationReceipt`, `eth_getUserOperationByHash`,
        `eth_gasPrice`, `eth_maxPriorityFeePerGas`.
      security:
        - aaProxyToken: []
      parameters:
        - $ref: "#/components/parameters/chainIdQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonRpcRequest"
            example:
              jsonrpc: "2.0"
              method: eth_sendUserOperation
              params: []
              id: 1
      responses:
        "200":
          description: Upstream bundler response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonRpcResponse"
        "400":
          description: Invalid JSON-RPC or method not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid AA proxy token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request too large (200 KB limit)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  retryAfterMs:
                    type: integer
...